==> analysis/01_Protein_read_data.rmd <==
---
title: "01_Protein_read_data"
author: "toobiwankenobi"
date: "2020-08-04"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction
This file will load the single-cell data and store it in an SingleCellExperiment data container. In order to successfully run this script, several .csv files have to provided in the data folder of this repository.

# Preparations

## Load libraries

First, we will load the libraries needed for this part of the analysis.

```{r load-libraries, message=FALSE}
library(data.table)
library(SingleCellExperiment)
```

## Read the data

```{r load data}
# load the single cell data
cells <- as.data.frame(fread(file = "data/data_for_analysis/protein/cell.csv",stringsAsFactors = FALSE))

# load the image level metadata
image_mat <- as.data.frame(read.csv(file = "data/data_for_analysis/protein/Image.csv",stringsAsFactors = FALSE))

# load the panel information
panel_mat <- read.csv(file = "data/data_for_analysis/protein/melanoma_1.06_protein.csv", sep= ",",  stringsAsFactors = FALSE )

# get an example file that contains the channel order
tags <- read.csv( "data/data_for_analysis/protein/20191021_ZTMA256.1_slide3_TH_s0_p9_r1_a1_ac_full.csv", header = FALSE)

# load acqusition meta data
acquisition_meta <- read.csv(file = "data/data_for_analysis/protein/acquisition_metadata.csv", stringsAsFactors = FALSE)

# load the clinical data. this is a clinical datatable that already contains the ImageNumbers. It has been prepared in the clinical metadata preparation.Rmd script (prepared for RNA and protein dataset separately)
clinical_mat <- read.csv("data/data_for_analysis/protein/clinical_data_protein.csv",stringsAsFactors = FALSE)
```

# Pre-processing

## Generate the counts data frame

```{r get count data }
cur_counts <- cells[,grepl("Intensity_MeanIntensityCorrected_FullStackFiltered",colnames(cells))]
```

## Get the scaling factor
the single cell data needs to be multiplied with the scaling factor (16 bit)

```{r scale the count data}
cur_counts <- cur_counts * image_mat$Scaling_FullStack[1]

# to order the channels according to channel number
channelNumber <- as.numeric(sub("^.*_c", "", colnames(cur_counts)))

cur_counts <- cur_counts[,order(channelNumber,decreasing = FALSE)]
```

## Get the information whether a cell is in the tumor mask

any cell that has more than 25 % of its Area in the tumor mask is considered as "TRUE" meaning inside the tumor.

```{r tumor mask}
tumor_mask <- cells$Intensity_MeanIntensity_tumormask * image_mat$Scaling_FullStack[1]
in_tumor <- tumor_mask > 0.25
```


## Prepare the cell-level metadata
this data frame contains the metadata for ever single cell and will later on be the colData in the single cell experiment object

the metadata will also have an entry called "Parent_nuclei" which holds information to the ObjectNumber of the Nuclei that the cell was derived from.
due to the down-scaling of the images some nuclei are lost and thus some cells do not have a Parent_nuclei

```{r column level metadata}
cell_meta <- DataFrame(CellNumber = cells$ObjectNumber,
                       ImageNumber = cells$ImageNumber,
                       Center_X = cells$Location_Center_X,
                       Center_Y = cells$Location_Center_Y,
                       Area = cells$AreaShape_Area,
                       MajorAxisLength = cells$AreaShape_MajorAxisLength,
                       MinorAxisLength = cells$AreaShape_MinorAxisLength,
                       NumberOfNeighbors = cells$Neighbors_NumberOfNeighbors_8,
                       Parent_nuclei = cells$Parent_nuclei,
                       in_tumor = in_tumor)


# add a unique cellID to each cell consisting of "dataset"+"ImageNumber"+"ObjectNumber"
cell_meta$cellID <- paste0("protein_",cell_meta$ImageNumber, "_",cell_meta$CellNumber)

rownames(cell_meta) <- cell_meta$cellID

# before we can add the clinical metadata to the cell_meta file we first have to make sure that the clinical metadata is also ordered according to the single cell data. In this case the reference for each single cell is the imagenubmer. so we need to order the clinical data by image number first.
clinical_mat <- clinical_mat[order(clinical_mat$ImageNumber),]
cell_meta <- cell_meta[order(cell_meta$ImageNumber),]
```

## Prepare the row-level metadata (panel/marker information)
here we prepare all the metadata for the rows in the single cell experiment object (rowData)

```{r row level metadata}
# the channel numbers are the rownumbers in the "tags" file that we create above
tags$channel <- as.numeric(rownames(tags))
colnames(tags) <- c("Metal.Tag","channel")

# include the channel information in the panel metadata (panel_mat)
panel_mat <- merge(panel_mat,tags,by="Metal.Tag")

# now we order the panel metadata by channel. therefore we first modify the column names
panel_mat <- panel_mat[order(panel_mat$channel,decreasing = FALSE),]

# we also revise the nomenclature of the clean targets to not contain special characters like "-" etc
panel_mat$clean_target
clean_target <- c("Vimentin", "Caveolin1", "HistoneH3", "SMA","CD15", "H3K27me3", "CD7","CXCR2", "HLADR", "S100", "CD19", "CD45RA", "Sox9", "TOX1", "CD20", "CD68",
                  "pERK", "CD3", "CD36", "p75", "PD1", "MiTF", "CD11b", "GrzB", "PDL1", "TCF7", "CD45RO", "FOXP3", "ICOS", "bCatenin", "CD8", "Collagen1", "Ki67",
                  "CD11c", "pS6", "CD4", "IDO1",  "SOX10", "CD303", "CD206", "PARP", "DNA1", "DNA2", "Ki67Pt198", "CD45", "MPO")
panel_mat$clean_target <- clean_target

rownames(panel_mat) <- panel_mat$clean_target
```

# Create SCE object

## Create the single cell experiment object

```{r create single cell object}
# create the SCE object
sce <- SingleCellExperiment(assays = list(counts = t(cur_counts)))

# Set marker name as rownames and cellID as colnames
rownames(sce) <- rownames(panel_mat)
colnames(sce) <- rownames(cell_meta)

# add the column and row metadata
colData(sce) <- cell_meta
rowData(sce) <- panel_mat

# asinh transformed counts as well as add the nuclear count data
assay(sce, "asinh") <- asinh(counts(sce))
```

## Assign the clinical data to the metadata slot

```{r}
# order according to ImageNumber
clinical_mat <- clinical_mat[order(clinical_mat$ImageNumber),]
metadata(sce) <- as.list(clinical_mat)
```

## Save the SCE object

```{r save sce object}
saveRDS(sce,file = "data/data_for_analysis/sce_protein.rds")
```
