==> analysis/02_Protein_annotations.rmd <==
---
title: "02_Protein_annotations"
author: "toobiwankenobi"
date: "2020-08-04"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Preparations
Here, we add more metadata and colData annotations to the SCE object which are needed for downstream analyses. 

## Load libraries

```{r, message=F}
library(SingleCellExperiment)
library(dplyr)
```

## Load data

```{r}
sce <- readRDS(file = "data/data_for_analysis/sce_protein.rds")
cur_sce <- data.frame(colData(sce))
cur_meta <- data.frame(metadata(sce))
```

# Add existing clinical colData

## Various variables

```{r}
cur_sce <- cur_sce %>%
  left_join(., cur_meta[,c("ImageNumber", "Location", "Adjuvant", "IHC_T_score", "Mutation", 
                           "Cancer_Stage", "Status_at_3m", "BlockID", "Description", "TissueType",
                           "MM_location", "Age_range", "Gender", "PatientID", "relapse")])

colData(sce)$Location <- cur_sce$Location
colData(sce)$Adjuvant <- cur_sce$Adjuvant
colData(sce)$E_I_D <- cur_sce$IHC_T_score
colData(sce)$Mutation <- cur_sce$Mutation
colData(sce)$Cancer_Stage <- cur_sce$Cancer_Stage
colData(sce)$Status_at_3m <- cur_sce$Status_at_3m
colData(sce)$BlockID <- cur_sce$BlockID
colData(sce)$Description <- cur_sce$Description
colData(sce)$TissueType <- cur_sce$TissueType
colData(sce)$MM_location <- cur_sce$MM_location
colData(sce)$Age <- cur_sce$Age_range
colData(sce)$Gender <- cur_sce$Gender
colData(sce)$PatientID <- cur_sce$PatientID
colData(sce)$relapse <- cur_sce$relapse
```

# Make clinical data interpretable for statistical analysis

## Simplified biopsy origin

```{r}
# create simplified location of biopsies
cur_sce$MM_location_simplified <- NA
cur_sce[grep("CTRL", cur_sce$Location), ]$MM_location_simplified <- "control"
cur_sce[grep("LN", cur_sce$MM_location), ]$MM_location_simplified <- "LN"
cur_sce[grep("skin", cur_sce$MM_location), ]$MM_location_simplified <- "skin"
cur_sce[is.na(cur_sce$MM_location_simplified) == TRUE, ]$MM_location_simplified <- "other"

# add to colData
colData(sce)$MM_location_simplified <- cur_sce$MM_location_simplified

# add to metadata
cur_meta <- left_join(cur_meta, distinct(cur_sce, MM_location_simplified, ImageNumber))
```

## Create therapy grouping before surgery

```{r}
# unique treatments
unique(cur_meta$Last_sys_treatment_before_surgery)

# group treatment types
cur_meta$treatment_group_before_surgery <- NA
cur_meta[cur_meta$MM_location_simplified == "control",]$treatment_group_before_surgery <- "control"
cur_meta[cur_meta$Last_sys_treatment_before_surgery %in% "untreated",]$treatment_group_before_surgery <- "untreated"
cur_meta[cur_meta$Last_sys_treatment_before_surgery %in% c("aPD1", "aPD1 + aCTLA4", "aCTLA4", 
                                                           "aPD1 + aCTLA4 or aPD1", "aPD1 + aLAG3"),]$treatment_group_before_surgery <- "ICI"
cur_meta[cur_meta$Last_sys_treatment_before_surgery %in% c("BRAFi + MEKi", "MEKi"),]$treatment_group_before_surgery <- "Targeted"
cur_meta[cur_meta$Last_sys_treatment_before_surgery %in% c("chemotherapy"),]$treatment_group_before_surgery <- "Chemotherapy"
cur_meta[is.na(cur_meta$Last_sys_treatment_before_surgery),]$treatment_group_before_surgery <- "unknown"

# add treatment type  to cur_sce
cur_sce <- left_join(cur_sce, cur_meta[,c("ImageNumber", "treatment_group_before_surgery")])

# add relapse to SCE
colData(sce)$treatment_group_before_surgery <- cur_sce$treatment_group_before_surgery
```

## Create therapy grouping after surgery

```{r}
# unique treatments
unique(cur_meta$Treatment_after_surgery)

# group treatment types
cur_meta$treatment_group_after_surgery <- NA
cur_meta[cur_meta$MM_location_simplified == "control",]$treatment_group_after_surgery <- "control"
cur_meta[cur_meta$Treatment_after_surgery %in% "untreated",]$treatment_group_after_surgery <- "untreated"
cur_meta[cur_meta$Treatment_after_surgery %in% c("aPD1 + aCTLA4", "aPD1", "aCTLA4", "aPD1 + aLAG3"),]$treatment_group_after_surgery <- "ICI"
cur_meta[cur_meta$Treatment_after_surgery %in% c("BRAFi + MEKi", "BRAFi", "MEKi", "BRAFi + MEKi +/- aPD1"),]$treatment_group_after_surgery <- "Targeted"
cur_meta[cur_meta$Treatment_after_surgery %in% c("chemotherapy"),]$treatment_group_after_surgery <- "Chemotherapy"
cur_meta[cur_meta$Treatment_after_surgery %in% c("TVEC"),]$treatment_group_after_surgery <- "TVEC"
cur_meta[cur_meta$Treatment_after_surgery %in% c("PC"),]$treatment_group_after_surgery <- "Palliative"
cur_meta[is.na(cur_meta$Treatment_after_surgery),]$treatment_group_after_surgery <- "unknown"

# add treatment type  to cur_sce
cur_sce <- left_join(cur_sce, cur_meta[,c("ImageNumber", "treatment_group_after_surgery")])

# add relapse to SCE
colData(sce)$treatment_group_after_surgery <- cur_sce$treatment_group_after_surgery
```

## Create binarized grouping based on number of treatments before surgery

```{r}
cur_meta$treatment_status_before_surgery <- NA
cur_meta[cur_meta$Location == "CTRL",]$treatment_status_before_surgery <- "control"
cur_meta[cur_meta$Location != "CTRL",]$treatment_status_before_surgery <- ifelse(cur_meta[cur_meta$Location != "CTRL",]$Nr_treatments_before_surgery == 0, "naive", "non-naive")

# add treatment type to cur_sce
cur_sce <- left_join(cur_sce, cur_meta[,c("ImageNumber", "treatment_status_before_surgery")])

# add to SCE
colData(sce)$treatment_status_before_surgery <- cur_sce$treatment_status_before_surgery
```

## Create Death grouping

```{r}
Date_death <- metadata(sce)$Date_death
sce$Date_death <- Date_death[sce$ImageNumber]

#binarize death
sce$Death <- "no"
sce[,is.na(sce$Date_death) == FALSE]$Death <- "yes"
```


# Add updated metadata to SCE object

```{r}
metadata(sce) <- as.list(cur_meta)
```

## Save the SCE object

```{r}
saveRDS(sce,file = "data/data_for_analysis/sce_protein.rds")
```
