==> analysis/04_1_Protein_celltype_classification.rmd <==
---
title: "04_1_Protein_celltype_classification"
author: "toobiwankenobi"
date: "2020-08-24"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Introduction

This script performs cell-type classification based on manually labelled cells. We will create increasing complexity for cell type labelling.

# Preparations

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Read in data 
First, we will read in the SingleCellExperiment object and load all libraries.

```{r read-sce, message=F}
library(caret)
library(scater)
library(tidyverse)
library(dittoSeq)
library(viridis)
library(doParallel)
library(ggpubr)
library(randomForest)
```

## Load data

```{r}
sce <- readRDS("data/data_for_analysis/sce_protein.rds")

# load all subsetted sce object from hierarchichal gating and combine the
label.files <- list.files("data/data_for_analysis/protein/celltype_classifier/", full.names = TRUE)

# Read in SCE objects
cur_sces <- lapply(label.files, readRDS)

# Merge SCE objects
# Due to relabelling, we now need to match the colData entries and remove rowData
cur_entries <- names(colData(cur_sces[[1]]))

cur_sces <- lapply(cur_sces, function(x){
  colData(x) <- colData(x)[,cur_entries]
  rowData(x) <- NA
  return(x)
  })
  
labelled_sce <- do.call("cbind", cur_sces)

# add rowData
rowData(labelled_sce) <- rowData(sce)
```

## Rename Celltypes

```{r}
unique(labelled_sce$cytomapper_CellLabel)

labelled_sce[,labelled_sce$cytomapper_CellLabel == "Bcell"]$cytomapper_CellLabel <- "B cell"
labelled_sce[,labelled_sce$cytomapper_CellLabel == "BnTcell"]$cytomapper_CellLabel <- "BnT cell"
labelled_sce[,labelled_sce$cytomapper_CellLabel == "Tcytotoxic"]$cytomapper_CellLabel <- "CD8+ T cell"
labelled_sce[,labelled_sce$cytomapper_CellLabel == "Thelper"]$cytomapper_CellLabel <- "CD4+ T cell"
labelled_sce[,labelled_sce$cytomapper_CellLabel == "Tregulatory"]$cytomapper_CellLabel <- "FOXP3+ T cell"
```


## Duplicates

```{r}
# how many duplicates do we have?
ncol(labelled_sce[,duplicated(labelled_sce$cellID) == T]) / ncol(labelled_sce[,duplicated(labelled_sce$cellID) == F]) * 100

# remove duplicates (more than 1 label per cellID)
unique_labels <- labelled_sce[,duplicated(labelled_sce$cellID) == F]
```

## Add labels to SCE object

```{r read_labels}
unique(unique_labels$cytomapper_CellLabel)

label_vector <- rep("unlabelled", ncol(sce))
names(label_vector) <- colnames(sce)
label_vector[colnames(unique_labels)] <- unique_labels$cytomapper_CellLabel

# add to sce
colData(sce)$layer_1_gated <- label_vector
```

## Create colour vector

Here, we will define a colour vector for the cell-types contained in layer 1.

```{r}
layer1_colours <- vector(length = length(unique(label_vector)))
names(layer1_colours) <- unique(label_vector)

layer1_colours["BnT cell"] <- "goldenrod2"
layer1_colours["B cell"] <- "green1"
layer1_colours["Macrophage"] <- "greenyellow"
layer1_colours["Neutrophil"] <- "blue1"
layer1_colours["CD4+ T cell"] <- "yellow"
layer1_colours["FOXP3+ T cell"] <- "sienna1"
layer1_colours["pDC"] <- "red2"
layer1_colours["Stroma"] <- "tomato"
layer1_colours["CD8+ T cell"] <- "deepskyblue"
layer1_colours["Tumor"] <- "sienna4"
layer1_colours["unknown"] <- "gray"
layer1_colours["unlabelled"] <- "gray"

# Save in SCE object
metadata(sce)$colour_vectors$layer_1 <- layer1_colours
```

## Quality control

In the next step, we will check the quality of the labels by:

1. checking how many cells contain multiple labels (see chunk 2)
2. how many cells of how many images are labeled
3. how balanced the classes are
4. if the selected cells actually express the markers that they are supposed to express


Next, we will check how many cells and how many images are labelled.

```{r quality-control-2}
# 2. How many cells of how many images are labelled

# Percent cells labelled
as_tibble(colData(sce)) %>% 
  summarise(labelled_cells = sum(layer_1_gated != "unlabelled")/n()) * 100

# Percent images labelled
as_tibble(colData(sce)) %>% 
  group_by(ImageNumber) %>%
  summarise(labelled_cells = sum(layer_1_gated != "unlabelled")) %>%
  ungroup() %>%
  summarise(labelled_images = sum(labelled_cells != 0)/n()) * 100


# Percent of cells labelled per image
as_tibble(colData(sce)) %>% 
  group_by(ImageNumber) %>%
  summarise(labelled_cells = sum(layer_1_gated != "unlabelled")/n(),
            number_cells = n()) %>%
  as.data.frame()
```

We will check how balanced the classes are across the images.

```{r quality-control-3}
# Total cells per class
as_tibble(colData(sce)) %>%
  group_by(layer_1_gated) %>%
  summarise(number_cells = n())

# Total cells per class and Sample
as_tibble(colData(sce)) %>%
  group_by(layer_1_gated, ImageNumber) %>%
  summarise(number_cells = n()) %>%
  as.data.frame() %>%
  head(.)
```

Now, we will check the expression of selected markers across the classes and visualize cell labels on UMAP. 

```{r quality-control-4}
lab_sce <- sce[,sce$layer_1_gated != "unlabelled"]
agr_sce <- aggregateAcrossCells(lab_sce, ids = colData(lab_sce)[,c("ImageNumber", "layer_1_gated")], 
                                average = TRUE)

assay(agr_sce, "asinh") <- asinh(counts(agr_sce))
assay(agr_sce, "scaled_asinh") <- t(scale(t(asinh(counts(agr_sce)))))

colnames(agr_sce) <- paste0(agr_sce$ImageNumber, "_", agr_sce$layer_1_gated)

# Define markers that were used for gating
rownames(sce)

cur_markers <- c("CD15", "S100", "CD20", "CD68", "CD3", "p75", "MiTF", "FOXP3", "CD8", "Collagen1",
                 "CD4", "IDO1", "SOX10", "CD303", "MPO")

# Non-scaled
dittoHeatmap(agr_sce[cur_markers,], assay = "asinh",
             cells.use = colnames(agr_sce[cur_markers,]),
            annot.by = c("ImageNumber", "layer_1_gated"), 
            order.by = "layer_1_gated", cluster_rows = FALSE,
            scale = "none", heatmap.colors = viridis(100), 
            annotation_colors = list(layer_1_gated = metadata(sce)$colour_vectors$layer_1))
```


```{r}
# Centered and scaled
dittoHeatmap(agr_sce[cur_markers,], assay = "scaled_asinh",
            annot.by = c("ImageNumber", "layer_1_gated"), 
            order.by = "layer_1_gated", cluster_rows = FALSE,
            annotation_colors = list(layer_1_gated = metadata(sce)$colour_vectors$layer_1),
            heatmap.colors = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            breaks = seq(-3, 3, length.out = 101))
```


```{r}
unlab_sce <- sce[,sce$layer_1_gated == "unlabelled"]

ggplot() +
  geom_point(aes(x = UMAP1, y = UMAP2, colour = layer_1_gated), 
             data = data.frame(UMAP1 = reducedDim(unlab_sce, "UMAP")[,1],
                               UMAP2 = reducedDim(unlab_sce, "UMAP")[,2],
                               layer_1_gated = colData(unlab_sce)$layer_1_gated)) +
    geom_point(aes(x = UMAP1, y = UMAP2, colour = layer_1_gated), size = 0.5, 
             data = data.frame(UMAP1 = reducedDim(lab_sce, "UMAP")[,1],
                               UMAP2 = reducedDim(lab_sce, "UMAP")[,2],
                               layer_1_gated = colData(lab_sce)$layer_1_gated)) +
  scale_color_manual(values = metadata(sce)$colour_vectors$layer_1) + 
  theme_bw()
```

# Random Forrest Training

After quality control, we will now use a random forest classifier to classify the remaining cells in the dataset. 

## Splitting by cell-types

In the first instance, we will split the labelled data based on their cell-types and ignore from which images the calls come.
In the current setting most images have been labelled but in the future we want to have a closer look at how well cells of non-labelled images are classified.

### Train and validate the classifier

We will first split the labelled data into training and test (validation) data at a ratio of 70/30 train/test.

```{r split-data}
set.seed(1234)
trainIndex <- createDataPartition(factor(lab_sce$layer_1_gated), p = 0.70)
train_sce <- lab_sce[,trainIndex$Resample1]
test_sce <- lab_sce[,-trainIndex$Resample1]
```

Here, we will first use a 10-fold crossvalidation by partitioning the data randomly across the full dataset.
This process is repeated 5 times.
We will also use parallel processing for time reasons.
For the `randomForrest` classifier, we need to tune the `mtry` parameter - the number of variables sampled for each split.

```{r train-model, message = FALSE}
# Define seeds for parallel processing
# Per iteration, we evaluate 10 models while tuning mtry
set.seed(222)
seeds <- vector(mode = "list", length = 11)
for (i in 1:10) {
  seeds[[i]] <- sample.int(5000, 10)
}

seeds[[11]] <- sample.int(5000, 1)

fitControl <- trainControl(method = "repeatedcv",
                           repeats = 1,
                           number = 10,
                           seeds = seeds)

cl <- makePSOCKcluster(round(detectCores()/2,0), outfile = "")
registerDoParallel(cl)

set.seed(1234)
start = Sys.time()
rffit <- train(x = t(assay(train_sce, "asinh")[rowData(sce)$good_marker,]), 
               y = factor(train_sce$layer_1_gated),
               method = "rf", ntree = 500,
               tuneLength = 10,
               trControl = fitControl,
               allowParallel = TRUE)
stopCluster(cl)
end = Sys.time()
print(end-start)
rffit
```

We will now have a look at the accuracy measures over iterations.
The only parameter that has been tuned is `mtry`.

```{r accuracy}
ggplot(rffit) + 
  geom_errorbar(data = rffit$results,
                aes(ymin = Accuracy - AccuracySD,
                    ymax = Accuracy + AccuracySD),
                width = 0.4)
```

We can also compute the confusion matrix:

```{r confusion-matrix}
confusionMatrix(rffit)
```

We will also look at the variable importance.

```{r variable-importance, fig.height = 15}
cur_varImp <- varImp(rffit)
plot(cur_varImp, top = 37)
```

Finally, we will validate the model using the test data.

```{r model-testing, dev=("pdf"), fig.width=4.3, fig.height=2.4}
cur_pred <- predict(rffit, newdata = t(assay(test_sce, "asinh")[rowData(sce)$good_marker,]))

cm <- confusionMatrix(data = cur_pred, reference = factor(test_sce$layer_1_gated))
cm

a <- data.frame(cm$byClass) %>%
  mutate(class = sub("Class: ", "", rownames(cm$byClass))) %>%
  ggplot() + 
  geom_point(aes(1 - Specificity, Sensitivity, 
                 size = Detection.Rate,
                 fill = class),
             shape = 21) + 
  scale_fill_manual(values = metadata(sce)$colour_vectors$layer_1) + 
  theme_bw() + 
  theme(text=element_text(size=12)) + 
  ylab("Sensitivity (TPR)") +
  xlab("1 - Specificity (FPR)")

legend <- get_legend(a)
a + theme(legend.position = "none")
```

```{r model-testing-legend, dev=("pdf"), fig.width=2.1, fig.height=5}
plot(legend)
```

We will also observe the distribution of classification probabilities per image and class:

```{r prediciton-probability, fig.width = 15}
cur_pred <- predict(rffit, newdata = t(assay(test_sce, "asinh")[rowData(sce)$good_marker,]), 
                    type = "prob")

cur_pred %>%
  mutate(class = test_sce$layer_1_gated,
         image = test_sce$ImageNumber) %>%
  reshape2::melt(id.vars = c("class", "image"), variable.name = "celltype", value.name = "probability") %>%
  filter(class == celltype) %>%
  ggplot() +
  geom_boxplot(aes(interaction(image), probability), outlier.size = 0.5) +
    facet_wrap(. ~ class) + 
    theme(axis.text.x = element_text(angle = 90, hjust = 1))
```

This plot shows the median probability for each image and class.

# Predicting new data

Finally, we will predict the labels of all other cells.
For cell-type classification, we will use the method that was trained across all images. 

```{r predict-cells}
start = Sys.time()
cell_labels.class <- as.character(predict.train(rffit, 
                       newdata = t(assay(unlab_sce[rowData(sce)$good_marker,], "asinh")), 
                       type = "raw"))
cell_labels.prob <- predict.train(rffit, 
                       newdata = t(assay(unlab_sce[rowData(sce)$good_marker,], "asinh")), 
                       type = "prob")
end = Sys.time()
print(end-start)
```

Store predictions in SCE object.
We will not overwrite the labels of the already labelled cells.

```{r store-predictions}
cell_labels <- sce$layer_1_gated
cell_labels[colnames(unlab_sce)] <- cell_labels.class

sce$celltype <- cell_labels
```

# Visualization

Here, we will visualize the predicted cell-types and their associated classification probabilities.


## Using reduced dimensions

```{r UMAP_celltypes_gated}
dittoDimPlot(sce[,sce$layer_1_gated != "unlabelled"], var = "celltype", reduction.use = "UMAP", size = 0.5, 
              color.panel = metadata(sce)$colour_vectors$layer_1, main = "Cell types gated")
```

```{r UMAP_celltypes_classified}
dittoDimPlot(sce[,sce$layer_1_gated == "unlabelled"], var = "celltype", reduction.use = "UMAP", size = 0.5, 
              color.panel = metadata(sce)$colour_vectors$layer_1, main = "Cell types classified") 
```

## Probabilities for all celltypes

```{r UMAP_probabilites_celltypes}
for (i in unique(cell_labels.class)) {
  cur_df <- data.frame(UMAP1 = reducedDim(unlab_sce, "UMAP")[,1],
       UMAP2 = reducedDim(unlab_sce, "UMAP")[,2],
       prob = cell_labels.prob[,i],
       class = cell_labels.class == i)
  
  ggplot() + geom_point(aes(UMAP1, UMAP2), data = cur_df[!cur_df$class,],
                       color = "gray") +
    geom_point(aes(UMAP1, UMAP2, color = prob), data = cur_df[cur_df$class,],
             size = 0.5)+
    scale_colour_viridis(name = paste0(i, " probability"))
}
```

## Visualization of marker expression

Finally, we will visualize the marker expression per cell type using the classified cells.

```{r heatmap-visualization}
unlab_sce <- sce[,sce$layer_1_gated == "unlabelled"]
agr_sce <- aggregateAcrossCells(unlab_sce, ids = colData(unlab_sce)[,c("ImageNumber", "celltype")], 
                                average = TRUE)
assay(agr_sce, "asinh") <- asinh(counts(agr_sce))

colnames(agr_sce) <- paste0(agr_sce$ImageNumber, "_", agr_sce$celltype)

rownames(sce)
# Non-scaled
dittoHeatmap(agr_sce[cur_markers,], assay = "asinh",
            annot.by = c("celltype"), 
            order.by = "celltype", cluster_rows = FALSE,
            scale = "none", heatmap.colors = viridis(100), 
            annotation_colors = list(celltype = metadata(sce)$colour_vectors$layer_1))

# Centered and scaled
dittoHeatmap(agr_sce[cur_markers,], assay = "asinh",
            annot.by = c("celltype"), 
            cluster_rows = FALSE,
            annotation_colors = list(celltype = metadata(sce)$colour_vectors$layer_1),
            heatmap.colors = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            breaks = seq(-3, 3, length.out = 101))
```

# Save RDS

```{r saveRDS}
saveRDS(sce, "data/data_for_analysis/sce_protein.rds")

# create data frame with class and probabilities and save as csv.
layer_1_dat <- as.data.frame(cell_labels.prob)
layer_1_dat$class <- cell_labels.class

#write.csv(layer_1_dat, file = "data/data_for_analysis/layer_1_classification_protein.csv")
```
