==> analysis/04_2_protein_classification_subclustering.rmd <==
---
title: "subluclustering and cell type definition"
author: "Daniel"
date: "19 12 2019"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# load packages

```{r load packages}
sapply(list.files("code/helper_functions/", full.names = TRUE), source)
library(LSD)
library(SingleCellExperiment)
library(ggplot2)
library(scater)
library(viridis)
library(igraph)
library(CATALYST)
library(workflowr)
```

# load data

```{r load data}
sce = readRDS(file = "data/data_for_analysis/sce_protein.rds")
```

```{r scatter plot to check cut offs for single markers}
for(i in 1:10){
scatter_x_y(sce,x = "CD8",y = "pERK",imagenumber =i, xlim = c(0,5), ylim = c(0,5))
  abline(h = 1.25)
  abline(v=1.5)
  abline(a = 0, b= 1)
}
```

## mean expression of markers per celltype

```{r heatmap mean per celltype asinh, fig.width= 15, fig.height=10}
# Aggregate the counts. calculateSummary is a function written from Nils
mean_sce <- calculateSummary(sce, split_by = "celltype", 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Linear scale 
plotHeatmap(mean_sce, exprs_values = "arcsinh", 
            features = c(rownames(mean_sce)), 
            colour_columns_by = "celltype", 
            color = viridis(100), cluster_rows = FALSE, gaps_row = 10)
```

## scaled mean expression of markers per celltype

```{r heatmap mean per celltype asinh scaled, fig.width= 10, fig.height=10}
# Linear scale 
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            features = c(rownames(mean_sce)), 
            colour_columns_by = "celltype", 
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-3,3), cluster_rows = TRUE, gaps_row = 10)
```

## definition of markers used for subclustering the super-classes

```{r subclustering of the celltypes} 
# Select markers
marker_list <- list()

marker_list$Tumor <- c("Ki67", "PDL1", "PARP", "H3K27me3","CXCR2","HLADR","S100","Sox9","pERK","CD36",
                       "p75","MiTF","bCatenin","Collagen1","pS6","IDO1","SOX10")

marker_list$Neutrophil <- c("CD15", "PDL1", "MPO", "CD11b")


marker_list$`FOXP3+ T cell` <- c("IDO1", "PD1", "FOXP3", "ICOS", "CD4", "Ki67", "GrzB",
                        "CD45RA", "CD45RO")

marker_list$`CD8+ T cell` <- c("IDO1", "TOX1", "PD1", "TCF7", "ICOS", "CD8","Ki67", "GrzB",
                        "CD45RA", "CD45RO")

marker_list$`CD4+ T cell` <- c("IDO1", "TOX1", "PD1", "TCF7", 
                        "FOXP3", "ICOS", "CD4", "Ki67", "GrzB",
                        "CD45RA", "CD45RO")

marker_list$`BnT cell` <- c("IDO1", "TOX1", "PD1", "TCF7", 
                        "FOXP3", "ICOS", "CD8", "CD4", "Ki67", "GrzB",
                        "CD45RA", "CD45RO", "CD20","HLADR","H3K27me3","pS6","pERK", "CD7")

marker_list$pDC <- c("CD303","IDO1","GrzB","CXCR2","CD11b")

marker_list$`B cell` <- c("Ki67","CD45RA", "CD45RO", "CD20","HLADR","H3K27me3","pS6","pERK")

marker_list$Macrophage <- c("Caveolin1", "PDL1", "PARP", "H3K27me3","CXCR2","HLADR","CD45RO","CD45RA","CD68","pERK","CD36", "Collagen1","CD11c","pS6","IDO1","CD206")

marker_list$Stroma <- c("Caveolin1", "CD68","CD36", "Collagen1","SMA")

marker_list$unknown <- rownames(sce[rowData(sce)$good_marker,])
```

## cluster the whole dataset with FlowSOM

FlowSOM first because it is faster

```{r sublcustering whole dataset using FlowSOM}
## the FlowSOM function from CATALYST needs an another column in the rowData of the sce to work properly:
rowData(sce)$marker_class <- "state"

# vector for clustering
fs_clustering <- vector(length = ncol(sce))

# create the "exprs" slot in the assay data (needed for CATALYST)
assay(sce, "exprs") <- assay(sce,"asinh")

# Macrophage, B cells, CD4+ T cell, CD8+ T cell, Tother and BnT cells will be clustered for a total of 6 clustes each
set.seed(12345)
for(i in c("Macrophage","CD4+ T cell","CD8+ T cell")){
  cur_sce <- sce[,sce$celltype == i]
  
  cur_sce <- CATALYST::cluster(cur_sce,features = marker_list[i][[1]], ydim = 2,xdim = 3,maxK = 4)
  
  fs_clustering[sce$celltype == i] <- cur_sce$cluster_id
}

# pDCs and Neutrophils will be clustered to 4 clusteres
for(i in c("pDC","Neutrophil","BnT cell","B cell","FOXP3+ T cell","Stroma", "unknown")){
  cur_sce <- sce[,sce$celltype == i]
  
  cur_sce <- CATALYST::cluster(cur_sce,features = marker_list[i][[1]],ydim = 2,xdim = 2,maxK = 3)
  
  fs_clustering[sce$celltype == i] <- cur_sce$cluster_id
}

# Tumor will be clustered into 9 clusters
cur_sce <- sce[,sce$celltype == "Tumor"]
  
cur_sce <- CATALYST::cluster(cur_sce,features = marker_list["Tumor"][[1]],ydim = 3,xdim = 3,maxK = 7)
  
fs_clustering[sce$celltype == "Tumor"] <- cur_sce$cluster_id

# Save in SCE object
colData(sce)$celltype_clustered <- as.factor(fs_clustering)

sce$celltype_clustered <- paste0(sce$celltype, "_", sce$celltype_clustered)
```

## visulalize clustering results for FlowSOM

```{r heatmap mean expression per celltype subclustered asinh, fig.width= 20, fig.height=15}
# Aggregate the counts
mean_sce <- calculateSummary(sce, split_by = c("celltype","celltype_clustered"), 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Linear scale 
plotHeatmap(mean_sce, exprs_values = "arcsinh",
            features = rownames(sce)[rowData(sce)$good_marker],
            colour_columns_by = c("celltype"), 
            color = viridis(100), 
            labels_col = mean_sce$celltype_clustered,
            show_colnames = TRUE, annotation_legend = FALSE, borders_color = NA)
```

```{r heatmap mean expression per celltype subclustered asinh scaled, fig.width= 20, fig.height=15}
# Scaled
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            features = rownames(sce)[rowData(sce)$good_marker],
            colour_columns_by = c("celltype"), 
            labels_col = mean_sce$celltype_clustered,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4, 4),legend = TRUE)

```

## numbers of cells per cluster

```{r counts per cluster}
table(sce$celltype_clustered)
```


## assign names to clusters

```{r names for clusters, fig.height= 15, fig.width=15, eval=FALSE}
annotations <- sce$celltype_clustered

# annotations[annotations == "other_4"] <- "Other_CD206Low_pS6Low"
# annotations[annotations == "other_2"] <- "Tumor_SOX9Low_SOX10+_MITF+_S100+_bCatenin++"
# annotations[annotations == "other_3"] <- "Other_SOX10Low_MITFLow_bCatenin+"
# annotations[annotations == "other_1"] <- "Other_CaveolinLow"
# annotations[annotations == "other_5"] <- "Vasculature_SMA++_Collagen+_Caveolin+"
# annotations[annotations == "other_6"] <- "Stroma_SMA+_Collagen++_Caveolin++_CD36++"
# 
# annotations[annotations == "Tumor_2"] <- "Tumor_SOX9++_p75+"
# annotations[annotations == "Tumor_3"] <- "Tumor_SOX10+_MITFLow_CD36Low"
# annotations[annotations == "Tumor_4"] <- "Tumor_SOX10Low_S100+"
# annotations[annotations == "Tumor_5"] <- "Tumor_SOX9++_SOX10+_MITF+_S100Low"
# annotations[annotations == "Tumor_8"] <- "Tumor_SOX9Low_SOX10Low_MITFLow_p75+"
# annotations[annotations == "Tumor_9"] <- "Tumor_SOX10+_MITF+_S100++"
# annotations[annotations == "Tumor_6"] <- "Tumor_S100+"
# annotations[annotations == "Tumor_1"] <- "Tumor_SOX10Low_MITFLow_HLADR+"
# annotations[annotations == "Tumor_7"] <- "Tumor_SOX10+_MITF++_S100+_pS6+"
# 
# annotations[annotations == "B cells_1"] <- "B cells_pErkLow_H3K27meLow"
# annotations[annotations == "B cells_4"] <- "B cells_pErkLow_H3K27meLow"
# annotations[annotations == "B cells_2"] <- "B cells_H3K27me_LowpS6Low"
# annotations[annotations == "B cells_3"] <- "B cells_CD19++_pErkLow_H3K27meLow"
# annotations[annotations == "B cells_5"] <- "B cells_Ki67++"
# annotations[annotations == "B cells_6"] <- "B cells_SMALow"
# 
# annotations[annotations == "BnT_1"] <- "B cells_CD3Low_CD4Low"
# annotations[annotations == "BnT_2"] <- "B cells_pErkLow_H3K27meLow"
# annotations[annotations == "BnT_6"] <- "BnT_CD8+"
# annotations[annotations == "BnT_4"] <- "BnT_CD8+"
# annotations[annotations == "BnT_5"] <- "BnT_CD4+_TCF7+"
# annotations[annotations == "BnT_3"] <- "BnT_CD4+_ICOS++_PD1++_Ki67+"
# 
# annotations[annotations == "T_other_4"] <- "Tcells_undefined"
# annotations[annotations == "T_other_5"] <- "Tcells_undefined"
# annotations[annotations == "T_other_3"] <- "Tcells_CD15_LowGrzB+"
# annotations[annotations == "T_other_1"] <- "Tcells_Ki67++"
# annotations[annotations == "T_other_2"] <- "Tcells_CaveolinLow_SMALow"
# annotations[annotations == "T_other_6"] <- "Tcells_undefined"
# 
# annotations[annotations == "T_helper_2"] <- "CD4+ T cell_CD11cLow_HLADRLow"
# annotations[annotations == "T_helper_5"] <- "CD4+ T cell_SMALow_CaveolinLow"
# annotations[annotations == "T_helper_1"] <- "CD4+ T cell_TCF7++"
# annotations[annotations == "T_helper_6"] <- "CD4+ T cell_FoxP3++_ICOS+_CD11cLow_CD206Low"
# annotations[annotations == "T_helper_3"] <- "CD4+ T cell_FoxP3Low_pERKLow"
# annotations[annotations == "T_helper_4"] <- "CD4+ T cell_FoxP3Low"
# 
# annotations[annotations == "T_cytotoxic_3"] <- "CD8+ T cell_PD1+_TOX1+_GrzBLow"
# annotations[annotations == "T_cytotoxic_5"] <- "CD8+ T cell_TCF7+_H3K27meLow"
# annotations[annotations == "T_cytotoxic_6"] <- "CD8+ T cell_PD1+_TOX1+_GrzB+_ICOSLow_Ki67+"
# annotations[annotations == "T_cytotoxic_1"] <- "CD8+ T cell_GrzBLow_SMALow_CaveolinLow"
# annotations[annotations == "T_cytotoxic_2"] <- "CD8+ T cell_GrzBLow_pS6Low"
# annotations[annotations == "T_cytotoxic_4"] <- "CD8+ T cell_PD1+_TOX1+_GrzBLow_ICOSLow_CD11cLow_CD206Low"
# 
# annotations[annotations == "Macrophage_4"] <- "Macrophage_CD68Low_CaveolinLow"
# annotations[annotations == "Macrophage_6"] <- "Tumor_CD68Low_S100Low_bCateninLow_MITFLow_SOX10Low_HLADR+"
# annotations[annotations == "Macrophage_2"] <- "Macrophage_CD68Low_CD206++_CD36Low"
# annotations[annotations == "Macrophage_1"] <- "Macrophage_CD68Low_pS6Low_S100Low"
# annotations[annotations == "Macrophage_3"] <- "Macrophage_CD68+_CD11c+_CD206++_CXCR2+_PDL1+"
# annotations[annotations == "Macrophage_5"] <- "Macrophage_CD68++_CD11c+_CD36+_Caveolin+_Collagen+"
# 
# annotations[annotations == "pDC_1"] <- "pDC_IDO1+_GrzB+"
# annotations[annotations == "pDC_2"] <- "pDC_IDO1+_TOX1+_CD11b+"
# annotations[annotations == "pDC_3"] <- "pDC_IDO1++_GrzB++"
# annotations[annotations == "pDC_4"] <- "pDC_Ido1+_PARP++"
# 
# annotations[annotations == "Neutrophil_2"] <- "Neutrophil_CD11b+_CD15++_MPO+_PDL1++"
# annotations[annotations == "Neutrophil_4"] <- "Neutrophil_CD11b+_CD15+_MPO+_PDL1Low"
# annotations[annotations == "Neutrophil_3"] <- "Neutrophil_CD11b++_CD15++_MPO++_CXCR2Low_PDL1+"
# annotations[annotations == "Neutrophil_1"] <- "Neutrophil_CD11b+_CD15+_MPO++"

sce$cellAnnotation <- annotations

mean_sce <- calculateSummary(sce, split_by = c("celltype","cellAnnotation"), 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            features = rownames(sce)[rowData(sce)$good_marker],
            colour_columns_by = c("celltype"), 
            labels_col = mean_sce$cellAnnotation,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4, 4),legend = TRUE)
```

__clustering does not really pick up PDL1 positive Tumor cells (which definitely exist). Potentially we have to investigate this with a manual cut off. arcsinh > 1.75 seems to be reasonable__

__Additionally, the clustering did not pick up proliferating and non prolifertating tumor clusters. this should also be investigated by manual cut offs. arcsinh > 1.25 seems reasonable.__

```{r counts per celltype, eval=FALSE}
celltype_counts <- sce$celltype

table(celltype_counts)
```


## save the single cell object with the clusters assigned to the cells

```{r save object with clusters}
# delete the "exprs" slot from the single cell experiment again.
assay(sce,"exprs") <- NULL

saveRDS(sce,file = "data/data_for_analysis/sce_protein.rds")
```
