==> analysis/04_2_RNA_classification_subclustering.rmd <==
---
title: "04_2_RNA_classification_subclustering"
author: "toobiwankenobi"
date: "2020-07-28"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Introduction

In this script we try to set global thresholds for some key markers (CD3 etc) and then sub-cluster the respective cells with markers that we know are expressed on those cells.

# Preparations

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Load libraries

```{r message=F}
sapply(list.files("code/helper_functions/", full.names = TRUE), source)
sapply(list.files("/Volumes/server_homes/daniels/Git/imcRtools/R/", full.names = TRUE), source)
library(LSD)
library(SingleCellExperiment)
library(ggplot2)
library(scater)
library(viridis)
library(igraph)
library(CATALYST)
library(workflowr)
library(gridExtra)
library(dplyr)
```

## Load data

```{r load data}
sce <- readRDS(file = "data/data_for_analysis/sce_RNA.rds")
```

# Marker for Clustering

## Define Superclasses for subclustering

```{r}
# store labels from randomForest separately
sce$celltype_rf <- sce$celltype

# assign superclasses
sce[,sce$celltype_rf %in% c("exhausted Tcytotoxic", "Tcytotoxic")]$celltype <- "CD8+ T cell"
sce[,sce$celltype_rf %in% c("CD134+ Tcell", "CD134- Tcell")]$celltype <- "CD8- T cell"
```

## Definition of markers used for subclustering the classified cells

```{r subclustering of the celltypes} 
marker_list <- list()

marker_list$Stroma <- c("SMA", "CK5", "Cadherin11", "FAP","pRB","B2M","GLUT1","T5_CCL4","T7_CCL18","CCR2","T1_CXCL8",
                        "T4_CXCL10","T8_CXCL13","T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9","T11_CCL8","CD31",
                        "T10_CCL19","cleavedPARP")

marker_list$Tumor <- c("HLADR", "S100","B2M","GLUT1","PDL1","SOX10","T5_CCL4","T7_CCL18","CCR2","T1_CXCL8","T4_CXCL10","T8_CXCL13",
                       "T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9","T11_CCL8","T10_CCL19","cleavedPARP",
                       "Mart1","pRB")

marker_list$Neutrophil <- c("CD15", "PDL1","GLUT1", "MPO","T5_CCL4","T7_CCL18","T1_CXCL8","T4_CXCL10","T8_CXCL13",
                       "T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9","T11_CCL8","T10_CCL19","cleavedPARP", "pRB")


marker_list$`CD8- T cell` <- c("CD3", "CD8", "PD1", "Lag3", "CCR2","CD38","PDL1", "HLADR", "CD134","T5_CCL4", "B2M", "CD31",
                       "T7_CCL18","T1_CXCL8","T4_CXCL10","T8_CXCL13","T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9",
                       "T11_CCL8","T10_CCL19","cleavedPARP", "pRB")

marker_list$`CD8+ T cell` <- c("CD3", "CD8", "PD1", "Lag3", "CCR2","CD38","PDL1", "HLADR", "CD134","T5_CCL4", "B2M", "CD31",
                       "T7_CCL18","T1_CXCL8","T4_CXCL10","T8_CXCL13","T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9",
                       "T11_CCL8","T10_CCL19","cleavedPARP", "pRB")


marker_list$Macrophage <- c("CD68", "CD163", "HLADR", "PDL1", "CD3", "CCR2","T5_CCL4","T7_CCL18","T1_CXCL8","T4_CXCL10","T8_CXCL13",
                       "T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9","T11_CCL8","T10_CCL19","cleavedPARP", "pRB")

marker_list$Vasculature <- c("CD31", "SMA","CK5","pRB", "Cadherin11","T5_CCL4","T7_CCL18","T1_CXCL8","T4_CXCL10","T8_CXCL13",
                       "T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9","T11_CCL8","T10_CCL19","cleavedPARP", "pRB")

marker_list$`HLA-DR` <- c("HLADR","CD68","CD3","CCR2","T5_CCL4","T7_CCL18","T1_CXCL8","T4_CXCL10","T8_CXCL13",
                       "T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9","T11_CCL8","T10_CCL19","cleavedPARP", "pRB")

marker_list$CD38 <- c("HLADR","CD38","CD3","CCR2","CD163","CD68","T5_CCL4","T7_CCL18","T1_CXCL8","T4_CXCL10","T8_CXCL13",
                       "T3_CXCL12","T12_CCL2","T2_CCL22","T9_CXCL9","T11_CCL8","T10_CCL19","cleavedPARP", "pRB")

marker_list$unknown <- rownames(sce[rowData(sce)$good_marker,])

# check if the names of the marker list match the names of the classified cells
all(names(marker_list) %in% unique(sce$celltype))
```

# Clustering

## Sub-cluster the whole dataset with FlowSOM

FlowSOM first because it is faster

```{r sublcustering whole dataset using FlowSOM}
## the FlowSOM function from CATALYST needs an another column in the rowData of the sce to work properly:
rowData(sce)$marker_class <- "state"

# vector for clustering
fs_clustering <- vector(length = ncol(sce))

# Macrophage, CD8- T cell, CD8+ T cell, Tumor will be clustered for a total of 6 clustes each
set.seed(12345)

# 6 clusters
for(i in c("CD8- T cell","Macrophage","CD8+ T cell","Tumor")){
  #subset cells for clustering
  cur_sce <- sce[,sce$celltype == i]
  names(assays(cur_sce)) <- c("counts", "exprs","scaled_counts", "scaled_asinh")
  cur_sce <- CATALYST::cluster(cur_sce, features = marker_list[i][[1]], ydim = 2, xdim = 3, maxK = 4)
  fs_clustering[sce$celltype == i] <- cur_sce$cluster_id
}

# 4 clusters
for(i in c("Neutrophil","CD38","Stroma","Vasculature","HLA-DR", "unknown")){
  cur_sce <- sce[,sce$celltype == i]
  names(assays(cur_sce)) <-  c("counts", "exprs","scaled_counts", "scaled_asinh")
  cur_sce <- CATALYST::cluster(cur_sce, features = marker_list[i][[1]], ydim = 2, xdim = 2, maxK = 3)
  fs_clustering[sce$celltype == i] <- cur_sce$cluster_id
}

# Save in SCE object
colData(sce)$celltype_clustered <- as.factor(fs_clustering)

sce$celltype_clustered <- paste0(sce$celltype, "_", sce$celltype_clustered)
```

# Visualization

## Visulalize clustering results for FlowSOM

```{r heatmap_mean_expression_per_celltypes_layer1_subclustered_asinh_scaled, fig.width= 20, fig.height=15}
# Aggregate the counts
mean_sce <- calculateSummary(sce, split_by = c("celltype_clustered", "celltype"), 
                             exprs_values = "counts")
# Exclude bad markers
mean_sce <- mean_sce[rowData(sce)$good_marker,]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Scaled
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            features = rownames(sce)[rowData(sce)$good_marker],
            colour_columns_by = "celltype", 
            labels_col = mean_sce$celltype_clustered,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4, 4),legend = TRUE)
```

## Numbers of cells per cluster

```{r counts per cluster}
celltype_counts <- sce$celltype_clustered

table(celltype_counts)
```

# Cluster Names

## Assign names to clusters

```{r names for clusters, fig.height= 15, fig.width=15}
annotations <- sce$celltype_clustered

# add annotation if wanted
#annotations[annotations == "Tumor_1"] <- "Tumor1_"
#annotations[annotations == "Tumor_2"] <- "Tumor2_"
#annotations[annotations == "Tumor_3"] <- "Tumor3_"
#annotations[annotations == "Tumor_4"] <- "Tumor4_"
#annotations[annotations == "Tumor_5"] <- "Tumor5_"
#annotations[annotations == "Tumor_6"] <- "Tumor6_"

sce$cellAnnotation <- annotations

mean_sce <- calculateSummary(sce, split_by = c("celltype_clustered","celltype"), 
                             exprs_values = "counts")

# Exclude bad markers
mean_sce <- mean_sce[rowData(sce)$good_marker,]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            features = rownames(sce)[rowData(sce)$good_marker],
            colour_columns_by = c("celltype"), 
            labels_col = mean_sce$celltype,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4, 4),legend = TRUE)
```

# Clustering Quality Control

## Marker expression density of clusters vs. manually gated - CD8+ T cell

```{r density_marker_Tcytotoxic, fig.width=15, fig.height=10}
# Density of Expression in different CD8+ T cell clusters vs. manually gated Lag3+ cells
cur_sce <- data.frame(colData(sce))
cur_exprs <- data.frame(t(assays(sce)[[2]]))
cur_exprs <- cbind(cur_exprs, cur_sce[,c("celltype", "celltype_rf", "celltype_clustered")])
cur_exprs$cellID <- rownames(cur_exprs)

cur_exprs <- cur_exprs %>%
  reshape2::melt(id.vars = c("cellID", "celltype", "celltype_rf", "celltype_clustered"), 
                 variable.name = "marker", value.name = "asinh")

# clustering
a <- cur_exprs %>%
  filter(celltype == "CD8+ T cell" & marker %in% marker_list$`CD8+ T cell`) %>%
  ggplot(data=., aes(asinh, color = celltype_clustered)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)

# manual gating
b <- cur_exprs %>%
  filter(celltype == "CD8+ T cell" & marker %in% marker_list$`CD8+ T cell`) %>%
  ggplot(data=., aes(asinh, color = celltype_rf)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)

grid.arrange(a, b, ncol = 2)
```


## CD8- T cell

```{r density_marker_Tcells, fig.width=15, fig.height=10}
# clustering
a <- cur_exprs %>%
  filter(celltype == "CD8- T cell" & marker %in% marker_list$`CD8- T cell`) %>%
  ggplot(data=., aes(asinh, color = celltype_clustered)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)

# manual gating
b <- cur_exprs %>%
  filter(celltype == "CD8- T cell" & marker %in% marker_list$`CD8- T cell`) %>%
  ggplot(data=., aes(asinh, color = celltype_rf)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)

grid.arrange(a, b, ncol = 2)
```

## Tumor

```{r density_marker_Tumor, fig.width=10, fig.height=10}
# clustering
cur_exprs %>%
  filter(celltype == "Tumor" & marker %in% marker_list$Tumor) %>%
  ggplot(data=., aes(asinh, color = celltype_clustered)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)
```

## Neutrophil

```{r density_marker_Neutrophil, fig.width=10, fig.height=10}
# clustering
cur_exprs %>%
  filter(celltype == "Neutrophil" & marker %in% marker_list$Neutrophil) %>%
  ggplot(data=., aes(asinh, color = celltype_clustered)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)
```

## Macrophage

```{r density_marker_Macrophage, fig.width=10, fig.height=10}
# clustering
cur_exprs %>%
  filter(celltype == "Macrophage" & marker %in% marker_list$Macrophage) %>%
  ggplot(data=., aes(asinh, color = celltype_clustered)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)
```

## CD38

```{r density_marker_CD38, fig.width=10, fig.height=10}
# clustering
cur_exprs %>%
  filter(celltype == "CD38" & marker %in% marker_list$CD38) %>%
  ggplot(data=., aes(asinh, color = celltype_clustered)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)
```

## unknown

```{r density_marker_unknown, fig.width=10, fig.height=10}
# clustering
cur_exprs %>%
  filter(celltype == "unknown" & marker %in% marker_list$unknown) %>%
  ggplot(data=., aes(asinh, color = celltype_clustered)) + 
  geom_density() +
  xlim(0,4) + 
  facet_wrap(~marker)
```

# SCE object

## Save the single cell object

```{r save object with clusters}
saveRDS(sce, "data/data_for_analysis/sce_RNA.rds")
```
