==> analysis/05_RNA_chemokine_expressing_cells.rmd <==
---
title: "05_RNA_chemokine_expressing_cells"
author: "Tobias Hoch"
date: "2020-07-28"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Introduction
Here, we define chemokine-expressing cells 

# Preparations

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Load libraries

```{r, message=F}
source("code/helper_functions/detect_mRNA_expression.R")
source("code/helper_functions/validityChecks.R")
library(SingleCellExperiment)
library(dplyr)
library(ggplot2)
library(scater)
library(CATALYST)
library(reshape2)
library(LSD)
library(data.table)
library(ComplexHeatmap)
library(corrplot)
library(pheatmap)
library(grid)
library(gridExtra)
library(tidyr)
library(colorRamps)
```

## Load the single cell experiment object and the image metadata

```{r}
sce <- readRDS(file = "data/data_for_analysis/sce_RNA.rds")
```

# Detect Chemokine Expressing Cells

## Detection of chemokine expressing cells

for the detection of chemokine expressing cells we make use of the fact that we also measured a negative control (DapB). 

```{r}
# get the names of the chemokine channels without the negative control channel
chemokine_channels = rownames(sce[which(grepl("T\\d+_",rownames(sce)) & ! grepl("DapB",rownames(sce))),])

# run function to define chemokine expressing cells 
output_list <- compute_difference(sce, 
                          cellID = "cellID", 
                          assay_name = "asinh", 
                          threshold = 0.01, 
                          mRNA_channels = chemokine_channels, 
                          negative_control = "T6_DapB", 
                          return_calc_metrics = TRUE)

# overwrite SCE object
sce <- output_list$output_sce
```

## Plot results from chemokine detection

```{r different_chemokine_detection_results,fig.height=12, fig.width=14}
# check difference between DapB and signal (histogram)
for(i in chemokine_channels){
  
  # subset whole data set for visualization purposes
  diff_chemo <- output_list[[i]]
  diff_chemo_sub <- diff_chemo[sample(nrow(diff_chemo), nrow(diff_chemo)*0.1), ]
  
  a = ggplot(data = diff_chemo_sub, aes(x=diff)) + 
  geom_histogram(binwidth = 0.01) + 
  xlab(paste("DapB mRNA signal subtracted from", i, sep = " ")) + ggtitle("Raw distribution") + 
    theme(plot.title = element_text(hjust = 0.5, size = 35), 
          axis.title = element_text(size = 25), 
          axis.text =  element_text(size = 15))

  b = ggplot(data = diff_chemo_sub, aes(x=scaled_diff)) + 
  geom_histogram(binwidth = 0.1, aes(fill = 
                                       ifelse(padj <= 0.01 & scaled_diff > 0, 'p<0.01', 'n.s.'))) + 
  xlab(paste("DapB mRNA signal subtracted from", i, sep = " ")) + ggtitle("Scaled distribution") + 
    labs(fill = "legend") +
    xlim(-5,7) +
    scale_fill_manual(values = c("black", "deepskyblue1")) +
        theme(plot.title = element_text(hjust = 0.5, size = 35), 
          axis.title = element_text(size = 25), 
          legend.text = element_text(size = 20), 
          legend.title = element_text(size=20),
          axis.text =  element_text(size = 15))
  
  # significant cells defined by subtraction
  c = ggplot(data=diff_chemo_sub, aes(x=mean_negative_control, y=mean_chemokine)) + 
    geom_point(alpha=0.2, aes(col = 
                                ifelse(padj <= 0.01 & scaled_diff > 0, 'p<0.01', 'n.s.'))) + 
    scale_color_manual(values = c("black", "deepskyblue1")) +
    xlim(0,5.5) + ylim(0,5.5) +
    ylab(paste("Mean expression of", i, sep=" ")) +
    xlab("Mean DapB mRNAexpression") +
    ggtitle(paste("DapB mRNA vs.", i, sep = " ")) + 
    theme(plot.title = element_text(hjust = 0.5, size = 35), 
          axis.title = element_text(size = 25), 
          legend.position = "none",
          axis.text =  element_text(size = 15))
  #png(file = paste("~/Daniel_volume/Rout_RNA/chemokine_detection_method_comparison/",i,"difference_distribution_BH_0.01.png", sep="_"), height = 1000, width = 1600)
  grid.arrange(a,b,c, nrow = 3, ncol=1)
  #dev.off()
}
```

# Basic Stats

## Basic numbers on the chemokine expressing cells 

```{r}
chemokines <- data.frame(colData(sce))
chemokines <- chemokines[, chemokine_channels]

# calculate the amount of cells that are positive for 1, 2 and multiple combinations. exclude column containing the ids (12)
single_combinations = chemokines[rowSums(chemokines[,-1]) == 1,-1]

# number of single chemokines positive cells 
nrow(single_combinations)

double_combinations = chemokines[rowSums(chemokines[,-1]) == 2,-1]
# number of single chemokines positive cells 
nrow(double_combinations)


multiple_combinations = chemokines[rowSums(chemokines[,-1]) >= 3,-1]
# number of cells that express 3 or more chemokines
nrow(multiple_combinations)


# number of double positives per chemokine
double_counts <- colSums(double_combinations)

# frequency matrix and corrplot for frequency matrix
double_combinations[double_combinations == 0] <- NA
count_matrix = psych::pairwiseCount(x=double_combinations)

# normalize the frequency matrix by the amount of double combinations that occur for each chemokine
frequency_matrix <- count_matrix
for (i in colnames(count_matrix)){
  frequency_matrix[,i] <- frequency_matrix[,i]/double_counts[i]
}
```

## Frequency double-positive cells
The next plot shows the frequencies of all double positive cell occurences. e.g. of all T4_CXCL10 expressing cells that also express another chemokine more than 50% express T9_CXCL9.

```{r frequency_of_double_positive_chemokines, fig.height=10, fig.width=10}
corrplot(frequency_matrix, is.corr = FALSE, tl.col = 'black', method = 'pie', type = 'full', 
         tl.srt = 45, tl.cex = 0.8, tl.offset = 0.5, cl.length = 2, cl.cex = 1, cl.align.text = "l", cl.ratio = 0.3,
         diag=TRUE, order = "hclust")
```

## Corrplot of Frequency matrix
Now we normalize the numbers of double positives by the numbers of all respective positive chemokines. this shows that usually between 20-40 percent of chemokine expressing cells are double positive expressors.

```{r freq_corrplot}
single_counts <- colSums(single_combinations)
frequency_matrix <- count_matrix
for (i in colnames(count_matrix)){
  frequency_matrix[,i] <- frequency_matrix[,i]/single_counts[i]
}

corrplot(frequency_matrix, is.corr = FALSE, tl.col = 'black', method = 'pie', type = 'full', 
         tl.srt = 45, tl.cex = 0.8, tl.offset = 0.5, cl.length = 2, cl.cex = 1, cl.align.text = "l", cl.ratio = 0.3,
         cl.lim = c(0,1), diag=TRUE, order = "hclust")
```

# SCE object

## Add data to SCE object

```{r add to sec object}
# general chemokine producer tag for every cell (logical binary)
sce$chemokine <- ifelse(rowSums(data.frame(colData(sce)[,chemokine_channels])) > 0, TRUE, FALSE)

# rename colData entry names
idx <- match(chemokine_channels, colnames(colData(sce)))
for(i in idx){
  colnames(colData(sce))[i] <- strsplit(colnames(colData(sce))[i], split = "_")[[1]][2]
}
```

## Add expressor info and colour_vector (with control samples)

```{r}
cur_df <- colData(sce)
cur_df <- as_tibble(cur_df)
cur_df <- cur_df[,grepl("CCL|CXCL|Dap",colnames(cur_df))]

for(i in colnames(cur_df)){
  cur_df[[i]] <- ifelse(cur_df[[i]]== 1,i,"NA")
}

cur_df <- cur_df %>%
  unite(expressor,sep = "_",na.rm =TRUE,remove=FALSE)

cur_df$expressor <- gsub("NA_","",cur_df$expressor)
cur_df$expressor <- gsub("_NA","",cur_df$expressor)

# summary table of all combinations
summary_cur_df <- table(cur_df$expressor)

# order table according to abundance of combinations
summary_cur_df<- summary_cur_df[order(-as.numeric(summary_cur_df))]

# combinations with more than 600 occurrences
targets <- names(which(summary_cur_df > 600))
targets <- targets[!targets == "NA"]

# add expressor info to sce object
sce$expressor <- cur_df$expressor

# add target names to metadata
metadata(sce)$chemokines_morethan600_withcontrol <- targets
```

## Save SCE object

```{r save sce object}
saveRDS(object = sce, file = "data/data_for_analysis/sce_RNA.rds")
```
