==> analysis/06_RNA_chemokine_patch_detection.rmd <==
---
title: "07_RNA_neighbourhood_analysis_chemokines"
author: "toobiwankenobi"
date: "2020-07-28"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

# Preparations

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Load libraries

```{r, message=F}
sapply(list.files("code/helper_functions", full.names = TRUE), source)
library(ComplexHeatmap)
library(SingleCellExperiment)
library(ggplot2)
library(ggrepel)
library(dplyr)
library(data.table)
library(stringr)
library(sf)
library(concaveman)
library(RANN)
```

## Load data

```{r load data}
sce = readRDS(file = "data/data_for_analysis/sce_RNA.rds")
```

# Chemokine Patch Detection and Analysis

## Detect Pachtes for Single Chemokines
Here we use our patch detection algorithm to detect patches of cells that express a certain chemokine. The following settings are used:
- We loop through all chemokines. If a cell produces mulitple chemokines, each chemokine is treated in a separate round of patch detection.
- min_clust_size: 1, even a single cell without producing neighbours is considered a patch. In the next chunk, we select only patches that at least contain 10 producing cells. 
- distance patch neighbours: 25µm
- expansion milieu: 30µm

```{r patch_milieu_detection}
start = Sys.time()
cur_sce <- data.frame(colData(sce))
# loop through all chemokines
for(i in names(cur_sce[,grepl(glob2rx("C*L*"),names(cur_sce))])) {

  # find clusters
  sce <- findPatch(input_sce = sce, 
                    IDs_of_interest = cur_sce[cur_sce[,names(cur_sce) == i] == 1,]$cellID, 
                    'cellID', 
                    'Center_X', 'Center_Y', 
                    'ImageNumber', 
                    distance = 25, 
                    min_clust_size = 1,
                    output_colname = paste(tolower(i), "only_clust", sep = ""))

  
  # define cells within/surrounding a cluster of chemokine producing cells
  sce <- findMilieu(sce, 
                     'cellID', 
                     'Center_X', 
                     'Center_Y', 
                     'ImageNumber', 
                     paste(tolower(i), "only_clust", sep = ""), 
                     distance = 30,
                     output_colname = paste(tolower(i), "only_comm", sep = ""))
} 
end = Sys.time()
print(end-start)
```

## Define pure clusters that contain more than 10 producing cells
Here we select patches which contain at least 10 producing cells. We then assign a to all milieu cells from that patch the milieu ID that was already assigned above.
All smaller patches get a 0. 

```{r bigger_patches}
cur_sce <- data.frame(colData(sce))

for(i in names(cur_sce[,grepl(glob2rx("*only_clust"),names(cur_sce))])) {
 
   # select cluster with more than 10 producing cells
  clust <- cur_sce %>%
    group_by_at(i) %>%
    summarise(n=n()) %>%
    distinct_at(i, .keep_all = TRUE) %>%
    filter(n>10)
  
  # remove cluster 0 (non-producing cells)
  clust <- clust[clust[,i] > 0,]
  
  # create new column
  name <- str_split(i, "_")[[1]][1]
  name <- paste0(name,"_pure")
  cur_sce[,name] <- 0
  
  # add cluster id to new column for all cells (producing AND non-producing) that are part of that community
  if(nrow(clust) > 0){
    colname_comm <- paste0(str_split(i, "_")[[1]][1],"_comm")
    cur_sce[cur_sce[,colname_comm] %in% clust[,i][[1]], name] <- as.numeric(cur_sce[cur_sce[,colname_comm] %in% clust[,i][[1]], colname_comm])
  }
}

# add to sce
sce$ccl4_pure <- cur_sce$ccl4only_pure
sce$ccl18_pure <- cur_sce$ccl18only_pure
sce$cxcl8_pure <- cur_sce$cxcl8only_pure
sce$cxcl10_pure <- cur_sce$cxcl10only_pure
sce$cxcl12_pure <- cur_sce$cxcl12only_pure
sce$cxcl13_pure <- cur_sce$cxcl13only_pure
sce$ccl2_pure <- cur_sce$ccl2only_pure
sce$ccl22_pure <- cur_sce$ccl22only_pure
sce$cxcl9_pure <- cur_sce$cxcl9only_pure
sce$ccl8_pure <- cur_sce$ccl8only_pure
sce$ccl19_pure <- cur_sce$ccl19only_pure
```

## Save SCE

```{r}
saveRDS(sce, file = "data/data_for_analysis/sce_RNA.rds")
```
