==> analysis/08_color_vectors.rmd <==
---
title: "08_coloring"
author: ""
date: "2020-11-25"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Introduction
Color schemes

# Preparations

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Load libraries

```{r, message=FALSE, warning=FALSE}
library(SingleCellExperiment)
library(colorRamps)
```

## Load data

```{r}
sce_rna <- readRDS("data/data_for_analysis/sce_RNA.rds")
sce_prot <- readRDS("data/data_for_analysis/sce_protein.rds")
```

## All celltypes

```{r}
colvec <- c("sienna4", "tomato", "gray65", "red3", "blueviolet", "deeppink", "blueviolet",
         "darkorchid1", "deepskyblue", "darkblue", "deepskyblue","aquamarine", "green",
         "darkgreen", "yellow")

names(colvec) <- c("Tumor", "Stroma", "unknown", "Vasculature", "CD8- T cell", "CD8+ T cell", "CD4+ T cell",
           "FOXP3+ T cell", "B cell", "BnT cell", "HLA-DR", "CD38" ,"Macrophage", "Neutrophil", "pDC")

pie(c(rep(1,length((colvec)))),col = colvec,labels = names(colvec))

col_celltypes <- colvec
names(col_celltypes) <- names(colvec)
```

# RNA

## Celltypes

```{r}
cell_rna <- col_celltypes[c("Tumor", "Stroma", "unknown", "Vasculature", "CD8- T cell", 
                            "CD8+ T cell", "HLA-DR", "CD38","Macrophage", "Neutrophil")]

metadata(sce_rna)$colour_vectors$celltype <- cell_rna
```

## Chemokines (Combinations)

```{r}
# add color vector to metadata
targets <- metadata(sce_rna)$chemokines_morethan600_withcontrol
color_chemo <- primary.colors(length(targets))
names(color_chemo) <- targets

#barplot(seq_along(targets), col=color_chemo, main="Pastel_hcl", names.arg = targets)
metadata(sce_rna)$colour_vectors$chemokine_combinations <- color_chemo
```

## Chemokines (single)

```{r}
col_vector_chemokines <- metadata(sce_rna)$colour_vector$chemokine_combinations
col_vector_chemokines <- col_vector_chemokines[c("CXCL13", "CXCL10", "CXCL9", "CCL2", "CXCL12", "CCL19", "CCL18", "CXCL8", "CCL4", "CCL22")]
col_vector_new_chemo <- c("forestgreen")
names(col_vector_new_chemo) <- c("CCL8")
  
col_vector_chemokines <- c(col_vector_chemokines, col_vector_new_chemo)

metadata(sce_rna)$colour_vectors$chemokine_single <- col_vector_chemokines
```

# Protein

## Celltypes

```{r}
cell_protein <- col_celltypes[c("Tumor", "Stroma", "unknown", "CD8+ T cell", 
                                "CD4+ T cell", "FOXP3+ T cell", "B cell", "BnT cell",
                                "Macrophage", "Neutrophil", "pDC")]

metadata(sce_prot)$colour_vectors$celltype <- cell_protein
```

## Save RDS

```{r}
saveRDS(sce_rna, "data/data_for_analysis/sce_RNA.rds")
saveRDS(sce_prot, "data/data_for_analysis/sce_protein.rds")
```
