==> analysis/08_RNA_celltypes_and_chemokine_expression_Tfrac_score.Rmd <==
---
title: "celltypes and chemokine expression split by Tfrac score"
author: "Daniel"
date: "18 08 2020"
output: html_document
---

## here we compare the fractions of cell type clusters and marker expressions for T cell score

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# load packages

```{r load packages, message=FALSE}
sapply(list.files("~/bbvolume/Git/imcRtools/R/", full.names = TRUE), source)
sapply(list.files("~/bbvolume/Git/ZTMA256_protein_data/code/helper_functions/", full.names = TRUE), source)
library(LSD)
library(SingleCellExperiment)
library(ggplot2)
library(scater)
library(viridis)
library(igraph)
library(CATALYST)
library(reshape2)
library(cowplot)
library(tidyverse)
```


# load data

```{r load data}
sce = readRDS(file = "../../bbvolume/server_homes/thoch/Git/melanoma_publication/data/sce_RNA.rds")

colour_vector <- readRDS("~/bbvolume/Git/ZTMA256_protein_data/data/colour_vector.rds")

# import T_frac_score and add to sce object
T_frac_score <- readRDS("data/T_frac_score_per_Description")

sce$T_frac_score_per_ImageNumber   <- T_frac_score[sce$Description]

# import also the T cell score per BlockID
T_frac_score <- readRDS("data/T_frac_score_per_BlockID")

sce$T_frac_score_per_BlockID   <- T_frac_score[sce$BlockID]
```


```{r adapt T cell score to numbers, fig.height=10, fig.width=10}

sce[,which(sce$T_frac_score_per_ImageNumber == "low" & sce$Location != "CTRL")]$T_frac_score_per_ImageNumber <- 0
sce[,which(sce$T_frac_score_per_ImageNumber == "med" & sce$Location != "CTRL")]$T_frac_score_per_ImageNumber <- 1
sce[,which(sce$T_frac_score_per_ImageNumber == "high" & sce$Location != "CTRL")]$T_frac_score_per_ImageNumber <- 2

  # subset sce to not contain CTRL samples
sce <- sce[,sce$Location != "CTRL"]
```



## celltypes per T_frac score
the fraction score defined on the protein dataset does also nicely split up the RNA data

```{r celltype fractions per T_frac_score_per_ImageNumber, fig.height=12, fig.width=20, message=FALSE}
p1 <- plotCellCounts(sce,colour_by = "celltype",split_by = "T_frac_score_per_ImageNumber", proportion = FALSE)
p2 <- plotCellCounts(sce,colour_by = "celltype",split_by = "T_frac_score_per_ImageNumber", proportion = TRUE)+theme(legend.position = "none")

p3 <- plotCellFracGroups(sce,CellClass = "celltype",color_by = "T_frac_score_per_ImageNumber")

bottom_row <- plot_grid(p1, p2, ncol = 2, rel_widths = c(1,1))
plot_grid(p3,bottom_row, ncol = 1, rel_heights = c(2,1))
```

__essentially no plasma cells (CD38+) in images with no cytotoxic T cells__

## producers and non-producers in community modules

```{r producers and non producers in community modules,fig.height=10, fig.width=20, message=FALSE}
p1 <- plotCellCounts(sce[,which(sce$community_module != 0 & sce$chemokine == TRUE)],colour_by = "celltype",split_by = "community_module", proportion = FALSE)
p2 <- plotCellCounts(sce[,which(sce$community_module != 0 & sce$chemokine == TRUE)],colour_by = "celltype",split_by = "community_module", proportion = TRUE)

p3 <- plotCellCounts(sce[,which(sce$community_module != 0 & sce$chemokine == FALSE)],colour_by = "celltype",split_by = "community_module", proportion = FALSE)
p4 <- plotCellCounts(sce[,which(sce$community_module != 0 & sce$chemokine == FALSE)],colour_by = "celltype",split_by = "community_module", proportion = TRUE)

plot_grid(p1,p2,p3,p4,ncol=2)
```

__this plot changed from the previous version. way less CD38 positive cells as producers.__

## Marker expression of celltypes in different community modules

here we check how the presence in a community affects the expression of different markers. 

```{r marker expression for community modules,fig.height=10, fig.width=20 }
# Aggregate the counts
mean_sce <- calculateSummary(sce, split_by = c( "community_module","celltype"), 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone|Vimentin", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Scaled
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            colour_columns_by = c("celltype", "community_module"), 
            labels_col = mean_sce$celltype,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-6, 6),legend = TRUE, cluster_cols = FALSE)

```
__community modules 1 and 4 drive the expression of PD1 and Lag3 in cytotoxic T cells. Modules 1, 3 and 8 drive the expression of CD134 in Thelper cells. Module 3 strongly drives the expression of cleaved PARP in tumor cells as well as the expression of GLUT1 in tumor cells, vasculature and stromal cells.__

## fraction celltypes for the pure communities

here we use the communities that were calculated by expansion only when a chemokine expressing cell of the same chemokine type was encountered within the specified distance.

```{r,fig.height=10,fig.width=15}
# first we generate a colData entry by which we can split the data
sce$pure_community <- ""

sce[,sce$cxcl9only_comm != 0]$pure_community <- "cxcl9only_comm"
sce[,sce$cxcl10only_comm != 0]$pure_community <- "cxcl10only_comm"
sce[,sce$cxcl13only_comm != 0]$pure_community <- "cxcl13only_comm"
sce[,sce$ccl2only_comm != 0]$pure_community <- "ccl2only_comm"
sce[,sce$ccl4only_comm != 0]$pure_community <- "ccl4only_comm"
sce[,sce$ccl18only_comm != 0]$pure_community <- "ccl18only_comm"
sce[,sce$ccl19only_comm != 0]$pure_community <- "ccl19only_comm"
sce[,sce$ccl22only_comm != 0]$pure_community <- "ccl22only_comm"
sce[,sce$cxcl8only_comm != 0]$pure_community <- "cxcl8only_comm"

p1 <- plotCellCounts(sce[,which(sce$pure_community != "" & sce$chemokine == TRUE)],colour_by = "celltype",split_by = "pure_community", proportion = FALSE)
p2 <- plotCellCounts(sce[,which(sce$pure_community != "" & sce$chemokine == TRUE)],colour_by = "celltype",split_by = "pure_community", proportion = TRUE)

p3 <- plotCellCounts(sce[,which(sce$pure_community != "" & sce$chemokine == FALSE)],colour_by = "celltype",split_by = "pure_community", proportion = FALSE)
p4 <- plotCellCounts(sce[,which(sce$pure_community != "" & sce$chemokine == FALSE)],colour_by = "celltype",split_by = "pure_community", proportion = TRUE)

plot_grid(p1,p2,p3,p4,ncol=2,labels = list("absolute count producers","proportional count producers","absolute counts non-producers","proportional count non-producers"))
```

__interestingly in the pure communities many more macrophages are now producers!__

## Marker expression of pure communities

```{r marker expression for pure community modules,fig.height=10, fig.width=20 }
# Aggregate the counts
mean_sce <- calculateSummary(sce, split_by = c( "pure_community","celltype"), 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone|Vimentin", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Scaled
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            colour_columns_by = c("celltype", "pure_community"), 
            labels_col = mean_sce$celltype,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4, 4),legend = TRUE, cluster_cols = FALSE)

```

__the pure community for CCL4 seems to induce the expression of Lag3 more strongly than the pure community of CXCL13. Botch those communities induce similar levels of PD1 expression__

```{r marker expression for Tcytotoxic within pure community modules,fig.height=10, fig.width=20 }
# Scaled
plotHeatmap(mean_sce[,mean_sce$celltype == "Tcytotoxic"], exprs_values = "arcsinh_scaled", 
            colour_columns_by = c("celltype", "pure_community"), 
            labels_col = mean_sce$celltype,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4, 4),legend = TRUE, cluster_cols = FALSE)

```
```{r marker expression for macrophages within pure community modules,fig.height=10, fig.width=20 }
# Scaled
plotHeatmap(mean_sce[,mean_sce$celltype == "Macrophage"], exprs_values = "arcsinh_scaled", 
            colour_columns_by = c("celltype", "pure_community"), 
            labels_col = mean_sce[,mean_sce$celltype == "Macrophage"]$celltype,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4, 4),legend = TRUE, cluster_cols = FALSE)

```

## pure clusters 

```{r,fig.height=10,fig.width=15}
# first we generate a colData entry by which we can split the data
sce$pure_cluster <- ""

sce[,sce$cxcl9only_clust != 0]$pure_cluster <- "cxcl9only_comm"
sce[,sce$cxcl10only_clust != 0]$pure_cluster <- "cxcl10only_clust"
sce[,sce$cxcl13only_clust != 0]$pure_cluster <- "cxcl13only_clust"
sce[,sce$ccl2only_clust != 0]$pure_cluster <- "ccl2only_clust"
sce[,sce$ccl4only_clust != 0]$pure_cluster <- "ccl4only_clust"
sce[,sce$ccl18only_clust != 0]$pure_cluster <- "ccl18only_clust"
sce[,sce$ccl19only_clust != 0]$pure_cluster <- "ccl19only_clust"
sce[,sce$ccl22only_clust != 0]$pure_cluster <- "ccl22only_clust"
sce[,sce$cxcl8only_clust != 0]$pure_cluster <- "cxcl8only_clust"

p1 <- plotCellCounts(sce[,sce$pure_cluster != ""],colour_by = "celltype",split_by = "pure_cluster", proportion = FALSE)
p2 <- plotCellCounts(sce[,sce$pure_cluster != ""],colour_by = "celltype",split_by = "pure_cluster", proportion = TRUE)

plot_grid(p1,p2)
```

```{r marker expression for pure clusters,fig.height=10, fig.width=20 }
# Aggregate the counts
mean_sce <- calculateSummary(sce, split_by = c( "pure_cluster","celltype"), 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone|Vimentin", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Scaled
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            colour_columns_by = c("celltype", "pure_cluster"), 
            labels_col = mean_sce$celltype,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-6,6),legend = TRUE, cluster_cols = FALSE)

```

__CCL22 pure cluster seems to strongly induce the expression of CD134 in Tcells and Tcytotoxic cells. CCL18 ist most strongly expressed in macrophages.__

## Tcytotoxic cluster fractions and marker profile for T_frac_score_per_BlockID

```{r Tcytotoxic cluster fractions for T_frac_score_per_BlockID, fig.height=10, fig.width=10, message=FALSE}
Date_death <- metadata(sce)$Date_death
sce$Date_death <- Date_death[sce$ImageNumber]

#binarize death
sce$Death <- "no"
sce[,grepl("20",sce$Date_death)]$Death <- "yes"


plotCellCounts(sce,colour_by = "pure_community",split_by = "T_frac_score_per_ImageNumber", proportion = FALSE)

plotCellFracGroups(sce,CellClass = "pure_community",color_by = "T_frac_score_per_ImageNumber")

plotCellFracGroups(sce[,sce$MM_location == "skin_subcutaneous"],CellClass = "pure_community",color_by = "Death")
```

## expression of markers in celltypes split by death within skin_subcutaneous

```{r, fig.width=10, fig.height=10}
# Aggregate the counts
mean_sce <- calculateSummary(sce[,sce$MM_location == "skin_subcutaneous"], split_by = c("Death", "celltype"), 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone|Vimentin", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Scaled
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            colour_columns_by = c("Death", "celltype"), 
            labels_col = mean_sce$celltype,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4,4),legend = TRUE, cluster_cols = FALSE)

```
## expression of markers in pure_communties split by death within skin_subcutaneous

```{r, fig.width=10, fig.height=10}
# Aggregate the counts
mean_sce <- calculateSummary(sce[,sce$MM_location == "skin_subcutaneous"], split_by = c("Death", "pure_community"), 
                             exprs_values = "counts")

# Exclude DNA and Histone
mean_sce <- mean_sce[!grepl("DNA|Histone|Vimentin", rownames(mean_sce)),]

# Transform and scale
assay(mean_sce, "arcsinh") <- asinh(assay(mean_sce, "meanCounts")) 
assay(mean_sce, "arcsinh_scaled") <- t(scale(t(asinh(assay(mean_sce, "meanCounts")))))

# Scaled
plotHeatmap(mean_sce, exprs_values = "arcsinh_scaled", 
            colour_columns_by = c("Death", "pure_community"), 
            labels_col = mean_sce$pure_community,
            show_colnames = TRUE, annotation_legend = TRUE, borders_color = NA,
            color = colorRampPalette(c("dark blue", "white", "dark red"))(100),
            zlim = c(-4,4),legend = TRUE, cluster_cols = FALSE)

```
