==> analysis/09_Tcell_Score.Rmd <==
---
title: "05 cytotoxic T cell quantification"
author: "Daniel"
date: "July 2020"
output: html_document
---

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Load libraries

```{r load packages}
sapply(list.files("code/helper_functions/", full.names = TRUE), source)
library(LSD)
library(SingleCellExperiment)
library(ggplot2)
library(scater)
library(viridis)
library(igraph)
library(CATALYST)
library(reshape2)
library(cowplot)
library(ggridges)
library(pheatmap)
library(tidyverse)
```

## Load Data

```{r load data, echo=FALSE}
sce_prot = readRDS(file = "data/data_for_analysis/sce_protein.rds")
sce_rna = readRDS(file = "data/data_for_analysis/sce_RNA.rds")

image_mat <- read.csv("data/data_for_analysis/protein/Image.csv")
image_mat_rna <- read.csv("data/data_for_analysis/rna/Image.csv")
```

# Cytotoxic T cell Scoring

## T Cell Density
We want to calculate the T cell density.
we calculate this per mm2. We use the count for cytotxic T cells as defined under “03_cell_type_definition” script

```{r}
cur_df <- data.frame(celltype = sce_prot$celltype,
                             ImageNumber = sce_prot$ImageNumber)

# count cell types per images
cellcount <- (t(table(cur_df)))

# here we get the imagesize from the image metadata
im_size <- (image_mat$Height_cellmask * image_mat$Width_cellmask)/1000000

# data frame
cellcount <- data.frame(cellcount)

# we calculate the density for each celltype for 1 mm2
cellcount$density <- cellcount$Freq/im_size[cellcount$ImageNumber]
cellcount <- cellcount[cellcount$celltype == "CD8+ T cell",]

# there are roughly 60 images with 50 or less cytotoxic T cells
hist(cellcount$density,breaks = 300)

# add cyotoxic T cell density to sce
cellcount$ImageNumber <- as.integer(cellcount$ImageNumber)
cur_sce <- data.frame(colData(sce_prot))
cur_sce <- left_join(cur_sce, cellcount[,c("ImageNumber", "density")])
sce_prot$cyotoxic_density_image <- cur_sce$density
```

## T cell density score per image
- absent: 1. quantile
- low: 2. quantile 
- med: 3. quantile
- high: 4. quantile

```{r}
# print quantiles
quantile(cellcount$density)
```

```{r}
# define a vector with all ImgeNumbers
T_density_scores <- c(1:length(unique(cellcount$ImageNumber)))

T_density_scores  <- rep("unassigned",length(unique(cellcount$ImageNumber)))

# use quantiles for scoring system
T_absent <- which(cellcount$density <= quantile(cellcount$density)[[2]])
T_low <- which(cellcount$density > quantile(cellcount$density)[[2]] & cellcount$density <= quantile(cellcount$density)[[3]])
T_med <- which(cellcount$density > quantile(cellcount$density)[[3]] & cellcount$density <= quantile(cellcount$density)[[4]])
T_high <- which(cellcount$density > quantile(cellcount$density)[[4]])

T_density_scores[T_absent] <- "absent"
T_density_scores[T_low] <- "low"
T_density_scores[T_med] <- "med"
T_density_scores[T_high] <- "high"

# now we add the information to the single cell experiment
sce_prot$Tcell_density_score_image <- T_density_scores[sce_prot$ImageNumber]
sce_prot$Tcell_density_score_image <- factor(sce_prot$Tcell_density_score_image, levels = c("absent", "low", "med", "high"))

# number of samples per group
data.frame(colData(sce_prot)) %>%
  distinct(Description, .keep_all = T) %>%
  group_by(Tcell_density_score_image) %>%
  summarise(n=n())
```

## E_I_D score compared to T_frac_score_per_ImageNumber

```{r check the cut offs}
cur_df <- data.frame(T_density = sce_prot$Tcell_density_score_image,
                     E_I_D = sce_prot$E_I_D)

table(cur_df)
```

## Add Scores to RNA data set

```{r export scores for RNA dataset}
sce_rna$infiltration <- NULL
sce_rna$T_frac_score_per_BlockID <- NULL
sce_rna$T_frac_score_per_ImageNumber <- NULL
sce_rna$T_frac_score_per_PatientID <- NULL
sce_rna$cyotoxic_density_image <- NULL

description_data <- data.frame(colData(sce_prot)) %>%
  distinct(Description, .keep_all = TRUE)

col_rna <- data.frame(colData(sce_rna))

# left_join
col_rna <- left_join(col_rna, description_data[,c("Description", "Tcell_density_score_image", "cyotoxic_density_image")])

# add to sce (attention: cytotoxic density is calculated on protein data set!)
sce_rna$Tcell_density_score_image <- col_rna$Tcell_density_score_image
sce_rna$cyotoxic_density_image <- col_rna$density
```

## Save updated SCE

```{r save sce}
saveRDS(sce_prot,file = "data/data_for_analysis/sce_protein.rds")
saveRDS(sce_rna,file = "data/data_for_analysis/sce_RNA.rds")
```
