==> analysis/10_Dysfunction_Score.rmd <==
---
title: "12_dysfunction_score"
author: ""
date: "2021-01-19"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Introduction
This script generates plots for Figure 4.

# Preparations

```{r setup}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

## Load libraries

First, we will load the libraries needed for this part of the analysis.

```{r, message=FALSE}
sapply(list.files("code/helper_functions", full.names = TRUE), source)
library(SingleCellExperiment)
library(reshape2)
library(tidyverse)
library(dplyr)
```

## Read the data

```{r}
sce_rna = readRDS(file = "data/data_for_analysis/sce_RNA.rds")
sce_prot = readRDS(file = "data/data_for_analysis/sce_protein.rds")

sce_rna$dysfunction_score <- NULL
sce_rna$dysfunction_density <- NULL
sce_prot$dysfunction_score <- NULL
sce_prot$dysfunction_density <- NULL
```

## Derive Scoring
We divide the images with a high CD8 cell density into two groups: Low Dysfunction (CD8+CXCL13+ < 5%) and High Dysfunction (CD8+CXCL13+ >= 5%). Margin punches from LN are not considered. 

```{r exhaustion scoring}
# paste density scoring and exhaustion scoring
dysfunction <- data.frame(colData(sce_rna)) %>%
  filter(Location != "CTRL") %>% # remove controls
  mutate(MM_location_punch = paste(MM_location, Location, sep = "_")) %>%
  filter(MM_location_punch != "LN_M") %>% # remove LN margin samples
  filter(Tcell_density_score_image %in% c("high")) %>%
  mutate(celltype2 = paste(celltype, CXCL13, sep = "_")) %>% # add CXCL13 info to celltype
  filter(celltype2 %in% c("CD8+ T cell_1", "CD8+ T cell_0")) %>%
  group_by(Description, celltype2) %>%
  summarise(n=n()) %>%
  reshape2::dcast(Description ~ celltype2, value.var = "n", fill = 0) %>%
  reshape2::melt(id.vars = c("Description"), variable.name = "celltype2", value.name = "n") %>%
  group_by(Description) %>%
  mutate(fraction = n / sum(n)) %>%
  filter(celltype2 == "CD8+ T cell_1") %>%
  ungroup() %>%
  mutate(dysfunction_score = ifelse(fraction >= median(fraction), "High Dysfunction", "Low Dysfunction")) %>%
  select(Description, dysfunction_score, fraction)

cur_rna <- data.frame(colData(sce_rna))
cur_rna <- left_join(cur_rna, dysfunction)
sce_rna$dysfunction_score <- cur_rna$dysfunction_score
sce_rna$dysfunction_density <- paste(sce_rna$Tcell_density_score_image, sce_rna$dysfunction_score, sep = " - ")

cur_prot <- data.frame(colData(sce_prot))
cur_prot <- left_join(cur_prot, dysfunction)
sce_prot$dysfunction_score <- cur_prot$dysfunction_score
sce_prot$dysfunction_density <- paste(sce_prot$Tcell_density_score_image, sce_prot$dysfunction_score, sep = " - ")
```

## Number of Patients in different groups

```{r}
data.frame(colData(sce_rna)) %>%
  filter(Location != "CTRL" & is.na(dysfunction_score) == FALSE) %>%
  distinct(PatientID, .keep_all = T) %>%
  group_by(dysfunction_score) %>%
  summarise(patients = n())
```

## Fraction of Dysfunctional Cells per Image

```{r}
ggplot(dysfunction, aes(x=reorder(Description,-fraction), y=fraction)) + 
  geom_bar(stat="identity") +
  geom_hline(yintercept = 0.05) +
  xlab("Description") +
  ylab("Fraction of CXCL13+ CD8+ T cells")

median(dysfunction$fraction)
```

## save SCE

```{r save sce}
saveRDS(sce_prot, file = "data/data_for_analysis/sce_protein.rds")
saveRDS(sce_rna, file = "data/data_for_analysis/sce_RNA.rds")
```
